Fix a race that could lead to orders out of sync in user data
The issue was in possibility to receive connect right after sending
new order request, thus new order won't be in pending_orders and
handler may complete fetching without it, while its executionReport
was missed during disconnect. While this event is extremely unlikely
it was still possible. The solution is to protect adding the pending
order with mutex that is acquired right before sending new order and
released right after adding order to the set of pending ones.